+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 
 ## Current Master
 
+## 3.4.0 2017-3-8
+
+- Add two most recently used custom RPCs to network dropdown menu.
+- Add personal_sign method support.
+- Add personal_ecRecover method support.
+- Add ability to customize gas and gasPrice on the transaction approval screen.
+- Increase default gas buffer to 1.5x estimated gas value.
+
 ## 3.3.0 2017-2-20
 
-- Add personal_sign and personal_ecRecover support.
 - net_version has been made synchronous.
 - Test suite for migrations expanded.
 - Network now changeable from lock screen.